Made lots of changes but some tests fail

+++ src/fsw/FCCode/EEPROMController.hpp
@@ -16,7 +16,7 @@ class EEPROMController : public TimedControlTask<void> {
      * @param offset
      * @param statefields
      */
-    EEPROMController(StateFieldRegistry& registry, unsigned int offset);
+    EEPROMController(StateFieldRegistry& registry, uint32_t offset);
 
     /**
      * @brief Gets the pointers to the statefields from the statefield
@@ -42,7 +42,7 @@ class EEPROMController : public TimedControlTask<void> {
      * @param position refers to the location of the statefield pointer, 
      * its address in the EEPROM, and its period in their respective vectors
      */
-    void update_EEPROM(unsigned int position);
+    void update_EEPROM(uint32_t position);
 
     /**
      * @brief Checks if the EEPROM is empty or not. The default value
@@ -52,10 +52,10 @@ class EEPROMController : public TimedControlTask<void> {
     bool check_empty();
 
     //the locations in the EEPROM in which the field values will be stored
-    std::vector<int> addresses;
+    std::vector<uint16_t> addresses;
 
     // Shared pointers to statefields that will be written to the EEPROM
-    std::vector<ReadableStateField<unsigned int>*> pointers;
+    std::vector<ReadableStateField<uint32_t>*> pointers;
 
     // Number of control cycles that must pass before the control task writes the value 
     // of a certain statefield to EEPROM
